Address #16 review: forward-proof `_distance`, reject k/nprobes 0
- Vector search now opts out of Lance's deprecated scoring autoprojection
  (`disable_scoring_autoprojection`) and builds the projection explicitly,
  always force-including `_distance` as a recognised system column. This
  survives the upcoming Lance behavior change instead of relying on the
  soon-to-be-removed legacy append. Naming `_distance` in `--columns` is now
  legal (deduplicated) rather than rejected by the projection resolver.
- `-k 0` and `--nprobes 0` are rejected at the clap layer with a clean message
  (`value_parser` range `1..`) instead of dying inside Lance / silently
  returning zero rows.
- Missing `--vector-file` now reports the offending path via a dedicated
  `VectorFileRead` error variant.

Tests: add clap-level rejection tests for `-k 0`, `--nprobes 0`, and a missing
query vector; add an adapter test that an explicit `_distance` projection is
not duplicated. All existing search tests remain green.

+++ src/cli.rs
@@ -232,11 +232,11 @@ pub enum Command {
         #[arg(long = "vector-file")]
         vector_file: Option<PathBuf>,
         /// Number of nearest neighbors to return.
-        #[arg(short = 'k', default_value_t = 10)]
-        k: usize,
+        #[arg(short = 'k', default_value_t = 10, value_parser = clap::value_parser!(u64).range(1..))]
+        k: u64,
         /// IVF partitions to probe (index tuning; no effect without an index).
-        #[arg(long)]
-        nprobes: Option<usize>,
+        #[arg(long, value_parser = clap::value_parser!(u64).range(1..))]
+        nprobes: Option<u64>,
         /// Refine factor for re-ranking search results (index tuning).
         #[arg(long = "refine-factor")]
         refine_factor: Option<u32>,
